fully address 8/9 PR comments
