+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-azure-sdk/readme.md
@@ -1,0 +1,25 @@
+# Project: Java Code Quality Analyzer
+
+## Project Description
+The Java Code Quality Analyzer, is a plugin designed to improve the quality of Java code. It provides an interactive tool window that offers real-time code suggestions. The following are proposed essential features, including rule set integration, telemetry connectivity, and Azure Toolkit integration.
+
+## Key Features
+
+- **Rule Set Integration**: This feature enables users to import and customize rule sets, tailoring the plugin to their specific needs.
+- **Telemetry Integration**: This feature connects the plugin to the backend with Application Insights, allowing for efficient data transmission.
+- **Azure Toolkit Integration**: This feature allows the plugin to integrate with the IntelliJ Azure toolkit, providing extended functionality.
+- **Editor Integration**: This feature offers continuous analysis and real-time code suggestions, enhancing the coding experience.
+- **Quick-Fix Actions**: This feature identifies issues and suggests quick actions within the tool window, facilitating immediate problem resolution.
+
+## User Interface
+
+- **Analysis Button**: Users can initiate code analysis which provides immediate feedback and sends telemetry information.
+- **Results Display**: This is a central area where users can view detailed violations and outcomes of the analysis.
+- **Rule Configuration Panel** : This space allows users to toggle rules for customization and disable telemetry if desired.
+- **Refresh/Clear Functionality**: Users can clear previous results and re-run the analysis at any time.
+
+## Rules
+1. #### Storage Upload without Length Check
+   Using Azure Storage upload APIs that don’t take a length parameter. This approach can lead to OutOfMemoryErrors, especially when dealing with large files or high-volume uploads
+    ##### Severity: INFO. 
+    #### Solution: Use APIs that take a length parameter.
