+++ PluginsAndFeatures/azure-toolkit-for-intellij/azure-intellij-plugin-azure-sdk/src/main/java/com/microsoft/azure/toolkit/intellij/azure/sdk/buildtool/ruleConfigs.json
@@ -1,0 +1,5 @@
+{
+  "StorageUploadWithoutLengthCheck": {
+    "methodsToCheck": ["uploadFromFile", "upload", "uploadWithResponse"]
+  }
+}
